refactor: split metadata to its own component
